Make dev PRs trigger the automated testing workflow (#70)
This pull request makes a small update to the automated testing workflow
configuration. The change ensures that tests are triggered not only for
pull requests targeting the `main` branch, but also for those targeting
the `dev` branch.

*
[`.github/workflows/automated-testing.yml`](diffhunk://#diff-7fcf76d142e9f61137571f967d394f00c735184feadb85ba245da503aae74e8dR7):
Added `dev` to the list of branches that trigger the workflow for pull
requests.